+++ crates/core/src/protocols/greedy_evalcheck/mod.rs
@@ -12,6 +12,7 @@
 //! protocols to reduce several evaluation claims to a single PCS opening per batch.
 
 mod error;
+mod logging;
 mod prove;
 mod verify;
