+++ crates/core/src/protocols/sumcheck/prove/mod.rs
@@ -5,6 +5,7 @@ mod batch_zerocheck;
 pub(crate) mod common;
 pub mod eq_ind;
 pub mod front_loaded;
+pub mod logging;
 pub mod oracles;
 pub mod prover_state;
 pub mod regular_sumcheck;
@@ -13,6 +14,7 @@ pub mod zerocheck;
 
 pub use batch_sumcheck::{batch_prove, SumcheckProver};
 pub use batch_zerocheck::{batch_prove as batch_prove_zerocheck, ZerocheckProver};
+pub use logging::PIOPCompilerFoldData;
 pub use oracles::{
 	constraint_set_sumcheck_prover, constraint_set_zerocheck_prover, split_constraint_set,
 };
